refactor: split apps/web/src/state/api.ts by domain (#585)
Pure refactor — no logic changes. The 1982-line `api.ts` is replaced by
a `state/api/` directory with one file per domain (activities, metrics,
locations, sync, settings, etc.) and a barrel `index.ts`. Existing
imports of `state/api` resolve through the index unchanged.

Largest file is now metrics.ts at 248 lines; all under the 500-line
target from issue #585.
