+++ apps/web/src/state/api/client.ts
@@ -1,0 +1,2 @@
+/** Browser's IANA timezone (e.g. "Europe/Stockholm"), sent to the backend for TZ-aware bucketing. */
+export const browserTz = Intl.DateTimeFormat().resolvedOptions().timeZone
