+++ apps/web/src/state/api/index.ts
@@ -1,0 +1,24 @@
+// Barrel re-export for the api module. Domain-split files; consumers import
+// `state/api` and resolve through this index.
+export * from './activities'
+export * from './activity-types'
+export * from './admin'
+export * from './audit'
+export * from './chart-data'
+export * from './correlations'
+export * from './custom-metrics'
+export * from './dashboard'
+export * from './deduction-rules'
+export * from './goals'
+export * from './locations'
+export * from './meals'
+export * from './metrics'
+export * from './notes'
+export * from './productivity'
+export * from './reports'
+export * from './screentime'
+export * from './settings'
+export * from './sync'
+export * from './training-load'
+export * from './trends'
+export * from './types'
